# frontend/tests/fixtures/openapi.json
{
  "components": {
    "schemas": {
      "AnomalyDto": {
        "additionalProperties": false,
        "description": "An anomaly plus the measurement it was based on.\n\nThe measurements are required: an anomaly row cannot outlive its telemetry\nrow (the foreign key cascades), and both are read in one transaction.",
        "properties": {
          "current_a": {
            "title": "Current A",
            "type": "number"
          },
          "detected_at": {
            "title": "Detected At",
            "type": "string"
          },
          "device_id": {
            "title": "Device Id",
            "type": "string"
          },
          "energy_wh": {
            "title": "Energy Wh",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "power_w": {
            "title": "Power W",
            "type": "number"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "telemetry_id": {
            "title": "Telemetry Id",
            "type": "integer"
          },
          "voltage_v": {
            "title": "Voltage V",
            "type": "number"
          }
        },
        "required": [
          "id",
          "telemetry_id",
          "device_id",
          "detected_at",
          "method",
          "score",
          "model_version",
          "reasons",
          "voltage_v",
          "current_a",
          "power_w",
          "energy_wh"
        ],
        "title": "AnomalyDto",
        "type": "object"
      },
      "AnomalyPageDto": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AnomalyDto"
            },
            "title": "Items",
            "type": "array"
          },
          "next_before_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Before Id"
          }
        },
        "required": [
          "items"
        ],
        "title": "AnomalyPageDto",
        "type": "object"
      },
      "AnomalySummary": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          }
        },
        "required": [
          "id",
          "method",
          "score",
          "model_version",
          "reasons"
        ],
        "title": "AnomalySummary",
        "type": "object"
      },
      "DeviceDto": {
        "additionalProperties": false,
        "properties": {
          "firmware_version": {
            "title": "Firmware Version",
            "type": "string"
          },
          "first_seen_at": {
            "title": "First Seen At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "last_seen_at": {
            "title": "Last Seen At",
            "type": "string"
          },
          "latest": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TelemetryDto"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "firmware_version",
          "status",
          "first_seen_at",
          "last_seen_at"
        ],
        "title": "DeviceDto",
        "type": "object"
      },
      "DeviceListDto": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DeviceDto"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "DeviceListDto",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HealthDto": {
        "additionalProperties": false,
        "properties": {
          "database": {
            "title": "Database",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "mqtt": {
            "title": "Mqtt",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "status",
          "database",
          "mqtt",
          "model",
          "version"
        ],
        "title": "HealthDto",
        "type": "object"
      },
      "TelemetryDto": {
        "additionalProperties": false,
        "properties": {
          "anomaly": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnomalySummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "boot_id": {
            "title": "Boot Id",
            "type": "string"
          },
          "current_a": {
            "title": "Current A",
            "type": "number"
          },
          "device_id": {
            "title": "Device Id",
            "type": "string"
          },
          "energy_wh": {
            "title": "Energy Wh",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "power_w": {
            "title": "Power W",
            "type": "number"
          },
          "received_at": {
            "title": "Received At",
            "type": "string"
          },
          "sampled_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sampled At"
          },
          "sensor_status": {
            "title": "Sensor Status",
            "type": "string"
          },
          "seq": {
            "title": "Seq",
            "type": "integer"
          },
          "voltage_v": {
            "title": "Voltage V",
            "type": "number"
          }
        },
        "required": [
          "id",
          "device_id",
          "boot_id",
          "seq",
          "sampled_at",
          "received_at",
          "voltage_v",
          "current_a",
          "power_w",
          "energy_wh",
          "sensor_status"
        ],
        "title": "TelemetryDto",
        "type": "object"
      },
      "TelemetryPageDto": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TelemetryDto"
            },
            "title": "Items",
            "type": "array"
          },
          "next_before_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Before Id"
          }
        },
        "required": [
          "items"
        ],
        "title": "TelemetryPageDto",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "AIoT PowerGuard",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/devices": {
      "get": {
        "operationId": "list_devices_api_v1_devices_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceListDto"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Devices",
        "tags": [
          "devices"
        ]
      }
    },
    "/api/v1/devices/{device_id}/anomalies": {
      "get": {
        "operationId": "anomaly_history_api_v1_devices__device_id__anomalies_get",
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "title": "Device Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 500,
              "maximum": 5000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "before_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnomalyPageDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Anomaly History",
        "tags": [
          "devices"
        ]
      }
    },
    "/api/v1/devices/{device_id}/latest": {
      "get": {
        "operationId": "latest_telemetry_api_v1_devices__device_id__latest_get",
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "title": "Device Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelemetryDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Latest Telemetry",
        "tags": [
          "devices"
        ]
      }
    },
    "/api/v1/devices/{device_id}/telemetry": {
      "get": {
        "operationId": "telemetry_history_api_v1_devices__device_id__telemetry_get",
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "title": "Device Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 500,
              "maximum": 5000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "before_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelemetryPageDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Telemetry History",
        "tags": [
          "devices"
        ]
      }
    },
    "/api/v1/health": {
      "get": {
        "operationId": "health_api_v1_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthDto"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health",
        "tags": [
          "health"
        ]
      }
    }
  }
}